fix(a2x): rename SDK package from a2x to @a2x/a2x for npm publish
- npm rejects unscoped 'a2x' as too similar to existing packages
- update all imports and workspace dependencies to @a2x/a2x
