+++ samples/express/package.json
@@ -9,7 +9,7 @@
     "start": "node dist/index.js"
   },
   "dependencies": {
-    "a2x": "workspace:*",
+    "@a2x/a2x": "workspace:*",
     "dotenv": "^16",
     "express": "^5.1.0"
   },
