+++ samples/nextjs/src/lib/a2x-setup.ts
@@ -7,8 +7,8 @@ import {
   A2XAgent,
   DefaultRequestHandler,
   OAuth2DeviceCodeAuthorization,
-} from "a2x";
-import { AnthropicProvider } from "a2x/anthropic";
+} from "@a2x/a2x";
+import { AnthropicProvider } from "@a2x/a2x/anthropic";
 import { globalTokens } from "@/app/oauth/token/route";
 
 const agent = new LlmAgent({
